+++ src/Faker/Internal.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE DeriveFunctor #-}
+{-# LANGUAGE BangPatterns #-}
 
 -- | Internal module
 module Faker.Internal
@@ -13,18 +14,15 @@ module Faker.Internal
   , resolver
   , refinedString
   , refinedText
-  , operateFields
-  , resolveFields
   , cachedRandomVec
   , cachedRandomUnresolvedVec
   , cachedRandomUnresolvedVecWithoutVector
   , cachedRegex
-  , interpolateNumbers
-  , interpolateString
   , resolveUnresolved
-  , uncons2
-  , operateField
   , modifyRandomGen
+  , resolveFields
+  , genericResolver
+  , genericResolver'
   ) where
 
 import Config
@@ -42,6 +40,9 @@ import Faker
 import Faker.Internal.Types (CacheFieldKey(..))
 import System.Random (StdGen, mkStdGen, randomR, split)
 import Text.StringRandom (stringRandom)
+import Fakedata.Parser
+import Data.Attoparsec.Text as P
+import Control.Monad (when)
 
 newtype Unresolved a = Unresolved
   { unresolvedField :: a
@@ -163,76 +164,55 @@ resolveUnresolved ::
   -> Unresolved (Vector Text)
   -> (FakerSettings -> Text -> m Text)
   -> m Text
-resolveUnresolved settings (Unresolved unres) resolverFn =
+resolveUnresolved settings (Unresolved unres) resolverFn = do
   let unresLen = V.length unres
       stdGen = getRandomGen settings
       (index, _) = randomR (0, unresLen - 1) stdGen
       randomItem = unres ! index
-      resolve =
-        if operateField randomItem "hello" == randomItem -- todo: remove hack
-          then pure $
-               interpolateString
-                 settings
-                 (interpolateNumbers settings randomItem)
-          else resolverFn settings randomItem
-   in if unresLen == 0
-        then throwM $ NoDataFound settings
-        else resolve
+  when (unresLen == 0) $ throwM $ NoDataFound settings
+  case P.parseOnly parseFakedata randomItem of
+       Left err -> throwM $ ParseError err
+       Right vals -> combineFakeIRValue settings resolverFn vals
 
-uncons2 :: Text -> Maybe (String, Text)
-uncons2 txt = do
-  (c1, rem1) <- T.uncons txt
-  (c2, rem2) <- T.uncons rem1
-  pure $ ((c1 : [c2]), rem2)
+resolveFakeIRValue :: (MonadIO m) => FakerSettings -> (FakerSettings -> Text -> m Text) -> (FakeIRValue,StdGen) -> m Text
+resolveFakeIRValue _ _ (Literal txt,_) = pure txt
+resolveFakeIRValue settings _ (Hash num,_) = pure $ resolveHash settings num
+resolveFakeIRValue settings _ (Ques num,_) = pure $ resolveQues settings num
+resolveFakeIRValue settings resolverFn (Resolve text,gen) = resolverFn (setRandomGen gen settings) text
 
--- operateField "#{hello} #{world}" "jam"
--- "jam #{world}"
-operateField :: Text -> Text -> Text
-operateField origWord word = helper origWord word []
-  where
-    helper :: Text -> Text -> String -> Text
-    helper txt word' acc =
-      case uncons2 txt of
-        Nothing -> origWord
-        Just (str, rem') ->
-          if str == "#{"
-            then let actualRem = dropTillBrace rem'
-                  in (T.pack acc) <> word' <> actualRem
-            else case T.uncons txt of
-                   Nothing -> origWord
-                   Just (c, rem2) -> helper rem2 word' (acc <> [c])
+combineFakeIRValue :: (MonadIO m) => FakerSettings -> (FakerSettings -> Text -> m Text) -> [FakeIRValue] -> m Text
+combineFakeIRValue settings resolverFn xs = do
+  vals <- mapM (resolveFakeIRValue settings resolverFn) (zip xs (stdgens (getRandomGen settings)))
+  pure $ T.concat vals
 
-operateFields :: Text -> [Text] -> Text
-operateFields origWord [] = origWord
-operateFields origWord (x:xs) = operateFields (operateField origWord x) xs
+resolveFields :: (MonadIO m, MonadThrow m) => Text -> m [FakeIRValue]
+resolveFields text = case P.parseOnly parseFakedata text of
+                       Left err -> throwM $ ParseError err
+                       Right vals -> pure vals
 
-dropTillBrace :: Text -> Text
-dropTillBrace txt = T.dropWhile (== '}') $ T.dropWhile (/= '}') txt
+genericResolver :: (MonadIO m, MonadThrow m) => FakerSettings -> Text -> (FakerSettings -> Text -> m Text) -> m Text
+genericResolver settings txt resolverFn = combineFakeIRValue settings resolverFn [Resolve txt]
 
--- λ> extractSingleField "#{hello} #{world} jam"
--- ("hello"," #{world} jam")
-extractSingleField :: Text -> (Text, Text)
-extractSingleField txt =
-  let (field, remaining) = T.span (\x -> x /= '}') txt''
-   in (T.drop 2 field, T.drop 1 remaining)
-  where
-    txt' = strip txt
-    txt'' = snd $ T.span (\x -> x /= '#') txt'
+genericResolver' :: (MonadIO m, MonadThrow m) => (FakerSettings -> Text -> m Text) -> FakerSettings -> Text -> m Text
+genericResolver' resolverFn settings txt = genericResolver settings txt resolverFn
 
--- λ> resolveFields "#{hello} #{world}"
--- ["hello","world"]
--- λ> resolveFields "#{hello} #{world} jam"
--- ["hello","world"]
-resolveFields :: Text -> [Text]
-resolveFields txt =
-  let (field, remaining) = extractSingleField txt
-      newField val =
-        if T.null val
-          then []
-          else [val]
-   in case T.null remaining of
-        True -> newField field
-        False -> newField field <> resolveFields remaining
+-- resolveHash settings 3
+-- "234"                                             
+resolveHash :: FakerSettings -> Int -> Text
+resolveHash settings num = T.pack $ helper settings num mempty
+    where
+      helper _ 0 acc = acc
+      helper settings !n acc = do
+        let (num, newGen) = randomR (0, 9) (getRandomGen settings)
+        helper (setRandomGen newGen settings) (n - 1) ((digitToChar num):acc)
+
+resolveQues :: FakerSettings -> Int -> Text
+resolveQues settings num = T.pack $ helper settings num mempty
+    where
+      helper _ 0 acc = acc
+      helper settings !n acc = do
+        let (char, newGen) = randomR ('A', 'Z') (getRandomGen settings)
+        helper (setRandomGen newGen settings) (n - 1) (char:acc)
 
 digitToChar :: Int -> Char
 digitToChar 0 = '0'
@@ -246,59 +226,6 @@ digitToChar 7 = '7'
 digitToChar 8 = '8'
 digitToChar 9 = '9'
 digitToChar x = error $ "Expected single digit number, but received " <> show x
-
-isHash :: Char -> Bool
-isHash c = c == '#'
-
--- >> interpolateNumbers "#####"
--- >> 23456
--- >> interpolateNumbers "ab-##"
--- >> ab-32
-interpolateNumbers :: FakerSettings -> Text -> Text
-interpolateNumbers fsettings txt = helper fsettings [] txt
-  where
-    helper settings acc text =
-      case T.null text of
-        True -> T.pack acc
-        False ->
-          case T.uncons text of
-            Nothing -> T.pack acc
-            Just (c, rem') ->
-              if isHash c
-                then let stdGen = getRandomGen settings
-                         (int, gen) = randomR (0, 9) stdGen
-                      in helper
-                           (setRandomGen gen settings)
-                           (acc ++ [digitToChar int])
-                           rem'
-                else helper settings (acc ++ [c]) rem'
-
-isQues :: Char -> Bool
-isQues c = c == '?'
-
--- >> interpolateString "?????"
--- >> ABCDE
--- >> interpolateString "32-##"
--- >> 32-ZF
-interpolateString :: FakerSettings -> Text -> Text
-interpolateString fsettings text = helper fsettings [] text
-  where
-    helper :: FakerSettings -> [Char] -> Text -> Text
-    helper settings acc txt =
-      case T.null txt of
-        True -> T.pack acc
-        False ->
-          case T.uncons txt of
-            Nothing -> T.pack acc
-            Just (c, remTxt) ->
-              if isQues c
-                then let stdGen = getRandomGen settings
-                         (int, gen) = randomR ('A', 'Z') stdGen
-                      in helper
-                           (setRandomGen gen settings)
-                           (acc ++ [int])
-                           remTxt
-                else helper settings (acc ++ [c]) remTxt
 
 resolver ::
      (MonadThrow m, MonadIO m)
@@ -359,6 +286,10 @@ insertToCache sdata field settings vec = do
   hmap <- getCacheField settings
   let hmap2 = HM.insert key vec hmap
   setCacheField hmap2 settings
+
+stdgens :: StdGen -> [StdGen]
+stdgens gen = let (g1, g2) = split gen
+              in [gen, g1, g2] <> (stdgens g2)
 
 -- TODO: Not efficient. Better to switch to splitmax once this gets
 -- resolved: https://github.com/phadej/splitmix/issues/23