+++ src/Faker/Provider/Job.hs
@@ -91,13 +91,7 @@ $(genParserUnresolved "job" "title")
 $(genProviderUnresolved "job" "title")
 
 resolveJobText :: (MonadIO m, MonadThrow m) => FakerSettings -> Text -> m Text
-resolveJobText settings txt = do
-  let fields = resolveFields txt
-  jobFields <-
-    mapM
-      (\(seed, field) -> resolveJobField (modifyRandomGen settings seed) field)
-      (zip [1 ..] fields)
-  pure $ operateFields txt jobFields
+resolveJobText = genericResolver' resolveJobField
 
 resolveJobField :: (MonadThrow m, MonadIO m) => FakerSettings -> Text -> m Text
 resolveJobField settings field@"seniority" =