+++ test/FakerSpec.hs
@@ -55,7 +55,7 @@ spec = do
   describe "Faker Compass Generate" $ do
     it "Resolver check" $ do
       ctries <- generate direction
-      (ctries) `shouldBe` "southwest"
+      (ctries) `shouldBe` "SE"
     it "Resolver check" $ do
       ctries <- generate $ listOf 5 direction
       (ctries) `shouldSatisfy` (\x -> Prelude.length x == 5)
@@ -115,7 +115,7 @@ spec = do
     describe "Restaurant" $ do
       it "name" $ do
         item <- generate FR.name
-        item `shouldBe` "Sweet House"
+        item `shouldBe` "Thirsty Pizza"
       it "name_suffix" $ do
         item <- generate FR.nameSuffix
         item `shouldSatisfy` isText